Fix host controller capabilities for Arasan SDHCI
The Arasan SD controller has a broken capabilities register, so the capabilities
need to be hardcoded in the driver. This updates the capabilities with the ones
from the official Linux in-kernel driver.

Also apply a bugfix to SetHighSpeed - apparently it's some timing issue.
Not sure if it actually applies to the Arasan, since the commit message only
mentions iProc Cygnus, but the Linux driver applies the fix for Arasan too.

With this, the Arasan's transfer rate should double to 25MB/s.